backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
django-environ>=0.11.0
django-celery-beat>=2.5.0
dj-database-url>=2.0.0
